Backend (Flask)/requirement.txt
bidict==0.22.0
click==8.1.3
dnspython==2.2.1
eventlet==0.33.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-SocketIO==5.2.0
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-engineio==4.3.2
python-socketio==5.6.0
six==1.16.0
Werkzeug==2.1.2
future==0.18.2
matplotlib==3.5.1
numpy==1.22.1
omegaconf==2.1.1
opencv_python_headless==4.5.5.62
pandas==1.4.0
Pillow==9.1.1
preprocessing==0.1.13
seaborn==0.11.2
setuptools==59.5.0
torch==1.11.0
torchmetrics==0.8.0
torchvision==0.12.0
tensorboard==2.8.0
protobuf==3.20.1
mediapipe==0.8.11